event_registration/requirements.txt
Django==3.2.9
django-crispy-forms==1.9.2
Pillow==8.4.0
django-bootstrap-datepicker-plus